Add wave_Computation.py
This is a very simple algorithm for
computing the solution to the one dimensional
wave equationnumerically. User inputs the
initial velocity and position, as well as
the boundaries, the position step size, and
the time step size, and the program will display
the numerical solution within the boundaries. Needs
to be updated with comments explaining the code, more
user-friendly parameters, and get rid of the
test case.