Add laplace_Computation.py file
Two  simple functions for numerically computing
solutions to the two dimensional laplace equation.
User enters initial matrix which represents the initial
and boundary conditions and the number of iterations they
wish the algorithm to proceed. Needs to be updated to
automatically calculate the number of iterations needed, and
comments need to be added to explain the code.